requirements.txt
# Web server
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0

# Twilio
twilio==9.0.0

# Speech-to-text
openai-whisper==20240930
faster-whisper==1.0.3

# LLM
ollama==0.3.0
httpx==0.27.0

# TTS
piper-tts==1.2.0

# Audio processing
numpy==1.26.4
sounddevice==0.4.7
soundfile==0.12.1
audioop-lts==0.2.1

# Utilities
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.4.0
